+++ dataquality-sampling/src/test/java/org/talend/dataquality/duplicating/TestConstants.java
@@ -1,0 +1,7 @@
+package org.talend.dataquality.duplicating;
+
+public class TestConstants {
+
+    public static final long RANDOM_SEED = 12345678L;
+
+}