+++ dataquality-statistics/src/main/java/org/talend/dataquality/statistics/frequency/recognition/DateTimePatternRecognizer.java
@@ -38,21 +38,6 @@ public class DateTimePatternRecognizer extends AbstractPatternRecognizer {
 
     private final SortedList<Map<Pattern, String>> frequentDatePatterns = new SortedList<>();
 
-    @Deprecated
-    public void addCustomDateTimePattern(String pattern) {
-        this.customDateTimePatterns.add(pattern);
-    }
-
-    @Deprecated
-    public void addCustomDateTimePatterns(List<String> patterns) {
-        this.customDateTimePatterns.addAll(patterns);
-    }
-
-    @Deprecated
-    public List<String> getCustomDateTimePattern() {
-        return customDateTimePatterns;
-    }
-
     @Override
     public RecognitionResult recognize(String stringToRecognize) {
         return recognize(stringToRecognize, DataTypeEnum.DATE);