+++ dataquality-statistics/src/test/java/org/talend/dataquality/statistics/datetime/SystemDateTimePatternManagerTest.java
@@ -64,34 +64,6 @@ public class SystemDateTimePatternManagerTest {
     }
 
     @Test
-    public void testIsMatchDateTimePatternInvalid() {
-        String pattern = "yyyy-MM-dd"; //$NON-NLS-1$
-        assertFalse(SystemDateTimePatternManager.isMatchDateTimePattern("2017-02-29", pattern, Locale.CHINESE)); //$NON-NLS-1$
-        assertFalse(SystemDateTimePatternManager.isMatchDateTimePattern("2017-04-31", pattern, Locale.US)); //$NON-NLS-1$
-        assertFalse(SystemDateTimePatternManager.isMatchDateTimePattern("31/11/2017", pattern, Locale.US)); //$NON-NLS-1$
-        assertFalse(SystemDateTimePatternManager.isMatchDateTimePattern("2015/01/32", pattern, Locale.US)); //$NON-NLS-1$
-        // the pattern is not adapt this value. should be "2011-05-02";
-        assertFalse(SystemDateTimePatternManager.isMatchDateTimePattern("2011-5-02", pattern, Locale.CHINESE)); //$NON-NLS-1$
-    }
-
-    @Test
-    public void testIsMatchDateTimePatternValid() {
-        String pattern = "yyyy-MM-dd"; //$NON-NLS-1$
-        assertTrue(SystemDateTimePatternManager.isMatchDateTimePattern("2016-02-29", pattern, Locale.CHINESE)); //$NON-NLS-1$
-        assertTrue(SystemDateTimePatternManager.isMatchDateTimePattern("2017-05-02", pattern, Locale.CHINESE)); //$NON-NLS-1$
-        assertTrue(SystemDateTimePatternManager.isMatchDateTimePattern("2017-11-30", pattern, Locale.CHINESE)); //$NON-NLS-1$
-        pattern = "dd/MM/yyy"; //$NON-NLS-1$
-        assertTrue(SystemDateTimePatternManager.isMatchDateTimePattern("30/01/2017", pattern, Locale.US)); //$NON-NLS-1$
-        pattern = "yyyy-MM-dd G"; //$NON-NLS-1$
-        assertFalse(SystemDateTimePatternManager.isMatchDateTimePattern("2017-02-15 AD", pattern, Locale.CHINESE)); //$NON-NLS-1$
-        assertFalse(SystemDateTimePatternManager.isMatchDateTimePattern("4714-11-12 BC", pattern, Locale.CHINESE)); //$NON-NLS-1$
-        pattern = "MMMM d, y GG";//$NON-NLS-1$
-        assertTrue(SystemDateTimePatternManager.isMatchDateTimePattern("March 15, 44 BC", pattern, Locale.US)); //$NON-NLS-1$
-        pattern = "MMMM d, u";//$NON-NLS-1$
-        assertTrue(SystemDateTimePatternManager.isMatchDateTimePattern("March 15, -43", pattern, Locale.US)); //$NON-NLS-1$
-    }
-
-    @Test
     public void testgetDateTimeFormatterByPattern() {
         DateTimeFormatter dateTimeFormatterByPattern = SystemDateTimePatternManager.getDateTimeFormatterByPattern("dd/MM/yyyy",
                 Locale.ENGLISH);