+++ dataquality-common/CHANGELOG.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - chore(TDQ-17710): Adopt the "Keep a Changelog" format for changelogs
 ### Changed
-N/A
+- chore(TDQ-17788): cancel the deprecation of setStoreInvalidValues in Quality Analyzers
 ### Removed
-N/A
+- chore(TDQ-17788): remove deprecated class HiraganaSmall
+- chore(TDQ-17788): remove unused Tests class
 ### Deprecated
 N/A
 ### Fixed