+++ dataquality-sampling/CHANGELOG.md
@@ -8,9 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - chore(TDQ-17710): Adopt the "Keep a Changelog" format for changelogs
 ### Changed
-N/A
+- chore(TDQ-17788): change RandomWrapper class to package protected
 ### Removed
-N/A
+- chore(TDQ-17788): remove unused AllDataqualitySamplingTests class
 ### Deprecated
 N/A
 ### Fixed